requirements.txt
Flask==3.0.3
Flask-Session==0.8.0
python-dotenv==1.0.1
openai==1.40.0
pandas==2.2.3
httpx==0.27.2
